Add a WIP signaled resource impl.
This will likely change. Currently dep signals are kept alive inside
the kernel arguments, however this isn't going to work for trying to
"hand off" a unique mut borrow baton (ie delegate the task of ensuring
a mutable borrow is unique to the command processor). So this interface
will need some work.